pyproject.toml
[project]
name = "mcp-powered-generative-ai-job-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.0.0",
    "google-genai>=1.43.0",
    "google-generativeai>=0.8.6",
    "pillow>=11.3.0",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.10",
    "streamlit>=1.49.1",
]
